requirements.txt
absl-py==2.1.0 ; python_full_version == "3.10.14"
astunparse==1.6.3 ; python_full_version == "3.10.14"
certifi==2024.8.30 ; python_full_version == "3.10.14"
charset-normalizer==3.3.2 ; python_full_version == "3.10.14"
cloudpickle==3.0.0 ; python_full_version == "3.10.14"
colorama==0.4.6 ; python_full_version == "3.10.14" and platform_system == "Windows"
contourpy==1.3.0 ; python_full_version == "3.10.14"
cycler==0.12.1 ; python_full_version == "3.10.14"
farama-notifications==0.0.4 ; python_full_version == "3.10.14"
flatbuffers==24.3.25 ; python_full_version == "3.10.14"
fonttools==4.53.1 ; python_full_version == "3.10.14"
gast==0.6.0 ; python_full_version == "3.10.14"
google-pasta==0.2.0 ; python_full_version == "3.10.14"
grpcio==1.66.1 ; python_full_version == "3.10.14"
gymnasium==0.29.1 ; python_full_version == "3.10.14"
h5py==3.11.0 ; python_full_version == "3.10.14"
idna==3.10 ; python_full_version == "3.10.14"
imageio==2.35.1 ; python_full_version == "3.10.14"
joblib==1.4.2 ; python_full_version == "3.10.14"
keras==3.5.0 ; python_full_version == "3.10.14"
kiwisolver==1.4.7 ; python_full_version == "3.10.14"
lazy-loader==0.4 ; python_full_version == "3.10.14"
libclang==18.1.1 ; python_full_version == "3.10.14"
markdown-it-py==3.0.0 ; python_full_version == "3.10.14"
markdown==3.7 ; python_full_version == "3.10.14"
markupsafe==2.1.5 ; python_full_version == "3.10.14"
matplotlib==3.9.2 ; python_full_version == "3.10.14"
mdurl==0.1.2 ; python_full_version == "3.10.14"
ml-dtypes==0.4.1 ; python_full_version == "3.10.14"
namex==0.0.8 ; python_full_version == "3.10.14"
networkx==3.3 ; python_full_version == "3.10.14"
numpy==1.24.4 ; python_full_version == "3.10.14"
opt-einsum==3.3.0 ; python_full_version == "3.10.14"
optree==0.12.1 ; python_full_version == "3.10.14"
packaging==24.1 ; python_full_version == "3.10.14"
pillow==10.4.0 ; python_full_version == "3.10.14"
protobuf==4.25.4 ; python_full_version == "3.10.14"
pygments==2.18.0 ; python_full_version == "3.10.14"
pyparsing==3.1.4 ; python_full_version == "3.10.14"
python-dateutil==2.9.0.post0 ; python_full_version == "3.10.14"
requests==2.32.3 ; python_full_version == "3.10.14"
rich==13.8.1 ; python_full_version == "3.10.14"
scikit-image==0.24.0 ; python_full_version == "3.10.14"
scikit-learn==1.5.2 ; python_full_version == "3.10.14"
scipy==1.14.1 ; python_full_version == "3.10.14"
setuptools==75.1.0 ; python_full_version == "3.10.14"
six==1.16.0 ; python_full_version == "3.10.14"
tensorboard-data-server==0.7.2 ; python_full_version == "3.10.14"
tensorboard==2.17.1 ; python_full_version == "3.10.14"
tensorflow-io-gcs-filesystem==0.37.1 ; python_full_version == "3.10.14"
tensorflow==2.17.0 ; python_full_version == "3.10.14"
termcolor==2.4.0 ; python_full_version == "3.10.14"
threadpoolctl==3.5.0 ; python_full_version == "3.10.14"
tifffile==2024.8.30 ; python_full_version == "3.10.14"
tqdm==4.66.5 ; python_full_version == "3.10.14"
typing-extensions==4.12.2 ; python_full_version == "3.10.14"
urllib3==2.2.3 ; python_full_version == "3.10.14"
werkzeug==3.0.4 ; python_full_version == "3.10.14"
wheel==0.44.0 ; python_full_version == "3.10.14"
wrapt==1.16.0 ; python_full_version == "3.10.14"
